requirements.txt
colormath==3.0.0
imageio==2.4.1
matplotlib==3.0.2
networkx==2.2
numpy==1.16.0
Pillow==5.4.1
